handle two-argument input, converting to legacy
instead of running with args like
myprog my first sentence . my second sentence .

we can run with
myprog "my first sentence ." "my second sentence >."

which is better for the CGI

this took some time because i had to debug
string->string splitting as opposed to
string->atom splitting

the legacy system expects a list of atoms, not a list of strings.

argh!